@@ -1,53 +1,39 @@
 {
-  "_args": [
-    [
-      "graphql@15.5.0",
-      "/home/matomi/Desktop/projects/GraphQL-Crash-Course/final/server"
-    ]
-  ],
-  "_from": "graphql@15.5.0",
-  "_id": "graphql@15.5.0",
-  "_inBundle": false,
-  "_integrity": "sha512-OmaM7y0kaK31NKG31q4YbD2beNYa6jBBKtMFT6gLYJljHLJr42IqJ8KX08u3Li/0ifzTU5HjmoOOrwa5BRLeDA==",
-  "_location": "/graphql",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "version",
-    "registry": true,
-    "raw": "graphql@15.5.0",
-    "name": "graphql",
-    "escapedName": "graphql",
-    "rawSpec": "15.5.0",
-    "saveSpec": null,
-    "fetchSpec": "15.5.0"
+  "name": "graphql",
+  "version": "16.12.0",
+  "description": "A Query Language and Runtime which can target any service.",
+  "license": "MIT",
+  "main": "index",
+  "module": "index.mjs",
+  "typesVersions": {
+    ">=4.1.0": {
+      "*": [
+        "*"
+      ]
+    },
+    "*": {
+      "*": [
+        "NotSupportedTSVersion.d.ts"
+      ]
+    }
   },
-  "_requiredBy": [
-    "/"
-  ],
-  "_resolved": "https://registry.npmjs.org/graphql/-/graphql-15.5.0.tgz",
-  "_spec": "15.5.0",
-  "_where": "/home/matomi/Desktop/projects/GraphQL-Crash-Course/final/server",
+  "sideEffects": false,
+  "homepage": "https://github.com/graphql/graphql-js",
   "bugs": {
     "url": "https://github.com/graphql/graphql-js/issues"
   },
-  "description": "A Query Language and Runtime which can target any service.",
-  "engines": {
-    "node": ">= 10.x"
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/graphql/graphql-js.git"
   },
-  "homepage": "https://github.com/graphql/graphql-js",
   "keywords": [
     "graphql",
     "graphql-js"
   ],
-  "license": "MIT",
-  "main": "index",
-  "module": "index.mjs",
-  "name": "graphql",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/graphql/graphql-js.git"
+  "engines": {
+    "node": "^12.22.0 || ^14.16.0 || ^16.0.0 || >=17.0.0"
   },
-  "sideEffects": false,
-  "types": "index.d.ts",
-  "version": "15.5.0"
+  "publishConfig": {
+    "tag": "latest"
+  }
 }
