@@ -1,168 +1,119 @@
 {
-  "_args": [
-    [
-      "uuid@8.3.2",
-      "/home/matomi/Desktop/projects/GraphQL-Crash-Course/final/server"
-    ]
+  "name": "uuid",
+  "version": "13.0.0",
+  "description": "RFC9562 UUIDs",
+  "type": "module",
+  "funding": [
+    "https://github.com/sponsors/broofa",
+    "https://github.com/sponsors/ctavan"
   ],
-  "_from": "uuid@8.3.2",
-  "_id": "uuid@8.3.2",
-  "_inBundle": false,
-  "_integrity": "sha512-+NYs2QeMWy+GWFOEm9xnn6HCDp0l7QBD7ml8zLUmJ+93Q5NF0NocErnwkTkXVFNiX3/fpC6afS8Dhb/gz7R7eg==",
-  "_location": "/uuid",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "version",
-    "registry": true,
-    "raw": "uuid@8.3.2",
-    "name": "uuid",
-    "escapedName": "uuid",
-    "rawSpec": "8.3.2",
-    "saveSpec": null,
-    "fetchSpec": "8.3.2"
-  },
-  "_requiredBy": [
-    "/",
-    "/apollo-server-core"
-  ],
-  "_resolved": "https://registry.npmjs.org/uuid/-/uuid-8.3.2.tgz",
-  "_spec": "8.3.2",
-  "_where": "/home/matomi/Desktop/projects/GraphQL-Crash-Course/final/server",
-  "bin": {
-    "uuid": "dist/bin/uuid"
-  },
-  "browser": {
-    "./dist/md5.js": "./dist/md5-browser.js",
-    "./dist/rng.js": "./dist/rng-browser.js",
-    "./dist/sha1.js": "./dist/sha1-browser.js",
-    "./dist/esm-node/index.js": "./dist/esm-browser/index.js"
-  },
-  "bugs": {
-    "url": "https://github.com/uuidjs/uuid/issues"
-  },
   "commitlint": {
     "extends": [
       "@commitlint/config-conventional"
     ]
   },
-  "description": "RFC4122 (v1, v4, and v5) UUIDs",
-  "devDependencies": {
-    "@babel/cli": "7.11.6",
-    "@babel/core": "7.11.6",
-    "@babel/preset-env": "7.11.5",
-    "@commitlint/cli": "11.0.0",
-    "@commitlint/config-conventional": "11.0.0",
-    "@rollup/plugin-node-resolve": "9.0.0",
-    "babel-eslint": "10.1.0",
-    "bundlewatch": "0.3.1",
-    "eslint": "7.10.0",
-    "eslint-config-prettier": "6.12.0",
-    "eslint-config-standard": "14.1.1",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-prettier": "3.1.4",
-    "eslint-plugin-promise": "4.2.1",
-    "eslint-plugin-standard": "4.0.1",
-    "husky": "4.3.0",
-    "jest": "25.5.4",
-    "lint-staged": "10.4.0",
-    "npm-run-all": "4.1.5",
-    "optional-dev-dependency": "2.0.1",
-    "prettier": "2.1.2",
-    "random-seed": "0.3.0",
-    "rollup": "2.28.2",
-    "rollup-plugin-terser": "7.0.2",
-    "runmd": "1.3.2",
-    "standard-version": "9.0.0"
+  "keywords": [
+    "uuid",
+    "guid",
+    "rfc4122",
+    "rfc9562"
+  ],
+  "license": "MIT",
+  "bin": {
+    "uuid": "./dist-node/bin/uuid"
   },
+  "sideEffects": false,
+  "types": "./dist/index.d.ts",
   "exports": {
     ".": {
-      "node": {
-        "module": "./dist/esm-node/index.js",
-        "require": "./dist/index.js",
-        "import": "./wrapper.mjs"
-      },
-      "default": "./dist/esm-browser/index.js"
+      "node": "./dist-node/index.js",
+      "default": "./dist/index.js"
     },
     "./package.json": "./package.json"
   },
   "files": [
-    "CHANGELOG.md",
-    "CONTRIBUTING.md",
-    "LICENSE.md",
-    "README.md",
     "dist",
-    "wrapper.mjs"
+    "dist-node",
+    "!**/test"
   ],
-  "homepage": "https://github.com/uuidjs/uuid#readme",
-  "husky": {
-    "hooks": {
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS",
-      "pre-commit": "lint-staged"
-    }
+  "devDependencies": {
+    "@babel/eslint-parser": "7.27.1",
+    "@commitlint/cli": "19.8.0",
+    "@commitlint/config-conventional": "19.8.0",
+    "@eslint/js": "9.26.0",
+    "bundlewatch": "0.4.1",
+    "commander": "13.1.0",
+    "eslint": "9.26.0",
+    "eslint-config-prettier": "10.1.2",
+    "eslint-plugin-prettier": "5.4.0",
+    "globals": "16.0.0",
+    "husky": "9.1.7",
+    "jest": "29.7.0",
+    "lint-staged": "15.5.2",
+    "neostandard": "0.12.1",
+    "npm-run-all": "4.1.5",
+    "prettier": "3.5.3",
+    "release-please": "17.0.0",
+    "runmd": "1.4.1",
+    "standard-version": "9.5.0",
+    "typescript": "5.2.2",
+    "typescript-eslint": "8.32.0"
   },
-  "keywords": [
-    "uuid",
-    "guid",
-    "rfc4122"
-  ],
-  "license": "MIT",
-  "lint-staged": {
-    "*.{js,jsx,json,md}": [
-      "prettier --write"
-    ],
-    "*.{js,jsx}": [
-      "eslint --fix"
-    ]
-  },
-  "main": "./dist/index.js",
-  "module": "./dist/esm-node/index.js",
-  "name": "uuid",
   "optionalDevDependencies": {
-    "@wdio/browserstack-service": "6.4.0",
-    "@wdio/cli": "6.4.0",
-    "@wdio/jasmine-framework": "6.4.0",
-    "@wdio/local-runner": "6.4.0",
-    "@wdio/spec-reporter": "6.4.0",
-    "@wdio/static-server-service": "6.4.0",
-    "@wdio/sync": "6.4.0"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/uuidjs/uuid.git"
+    "@wdio/browserstack-service": "9.2.1",
+    "@wdio/cli": "9.2.1",
+    "@wdio/jasmine-framework": "9.2.1",
+    "@wdio/local-runner": "9.2.1",
+    "@wdio/spec-reporter": "9.1.3",
+    "@wdio/static-server-service": "9.1.3"
   },
   "scripts": {
     "build": "./scripts/build.sh",
+    "build:watch": "tsc --watch -p tsconfig.json",
     "bundlewatch": "npm run pretest:browser && bundlewatch --config bundlewatch.config.json",
-    "docs": "( node --version | grep -q 'v12' ) && ( npm run build && runmd --output=README.md README_js.md )",
-    "docs:diff": "npm run docs && git diff --quiet README.md",
-    "eslint:check": "eslint src/ test/ examples/ *.js",
-    "eslint:fix": "eslint --fix src/ test/ examples/ *.js",
-    "examples:browser:rollup:build": "cd examples/browser-rollup && npm install && npm run build",
-    "examples:browser:webpack:build": "cd examples/browser-webpack && npm install && npm run build",
-    "examples:node:commonjs:test": "cd examples/node-commonjs && npm install && npm test",
-    "examples:node:esmodules:test": "cd examples/node-esmodules && npm install && npm test",
+    "docs:diff": "npm run docs && git diff --quiet -I \"[0-9a-f-]{36}\" README.md",
+    "docs": "npm run build && npx runmd --output=README.md README_js.md",
+    "eslint:check": "eslint src/ test/ examples/ *.[jt]s",
+    "eslint:fix": "eslint --fix src/ test/ examples/ *.[jt]s",
+    "examples:browser:rollup:build": "cd examples/browser-rollup && npm run build",
+    "examples:browser:webpack:build": "cd examples/browser-webpack && npm run build",
+    "examples:node:esmodules:test": "cd examples/node-esmodules && npm test",
+    "examples:node:jest:test": "cd examples/node-jest && npm test",
+    "examples:node:typescript:test": "cd examples/typescript && npm test",
     "lint": "npm run eslint:check && npm run prettier:check",
     "md": "runmd --watch --output=README.md README_js.md",
-    "prepack": "npm run build",
-    "pretest": "[ -n $CI ] || npm run build",
+    "prepack": "npm run build -- --no-pack",
+    "prepare": "husky",
+    "prepublishOnly": "npm run build",
     "pretest:benchmark": "npm run build",
-    "pretest:browser": "optional-dev-dependency && npm run build && npm-run-all --parallel examples:browser:**",
+    "pretest:browser": "./scripts/iodd && npm run build && npm-run-all --parallel examples:browser:**",
     "pretest:node": "npm run build",
-    "prettier:check": "prettier --ignore-path .prettierignore --check '**/*.{js,jsx,json,md}'",
-    "prettier:fix": "prettier --ignore-path .prettierignore --write '**/*.{js,jsx,json,md}'",
+    "pretest": "npm run build",
+    "prettier:check": "prettier --check .",
+    "prettier:fix": "prettier --write .",
     "release": "standard-version --no-verify",
-    "test": "BABEL_ENV=commonjs node --throw-deprecation node_modules/.bin/jest test/unit/",
-    "test:benchmark": "cd examples/benchmark && npm install && npm test",
+    "test:benchmark": "cd examples/benchmark && npm test",
     "test:browser": "wdio run ./wdio.conf.js",
     "test:node": "npm-run-all --parallel examples:node:**",
-    "test:pack": "./scripts/testpack.sh"
+    "test:watch": "node --test --enable-source-maps --watch dist-node/test/*.js",
+    "test": "node --test --enable-source-maps dist-node/test/*.js"
   },
-  "sideEffects": false,
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/uuidjs/uuid.git"
+  },
+  "lint-staged": {
+    "*": [
+      "prettier --no-error-on-unmatched-pattern --write"
+    ],
+    "*.{js,jsx}": [
+      "eslint --no-error-on-unmatched-pattern --fix"
+    ]
+  },
   "standard-version": {
     "scripts": {
       "postchangelog": "prettier --write CHANGELOG.md"
     }
   },
-  "version": "8.3.2"
+  "packageManager": "npm@11.3.0"
 }
