@@ -7,4 +7,4 @@ const Category = {
     }
 }
 
-module.exports = Category
+export default Category
