@@ -16,5 +16,5 @@ const Query = {
     }
 }
 
-module.exports = Query
+export default Query
 
